backend/src/tailored_resume_generator/requirements.txt
langchain_core
python-dotenv
langchain_openai
langchain_anthropic
boto3